fix: Scan book will return the ISBN to the text box instead of adding it directly to the library.
Scan ISBN was adding the book directly to the library. However, it is more consistent to return the ISBN as a string to the text box in the add book view. This gives the user a preview of the book he has just scanned for him to decide whether or not to add to the library